+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Lua
   related to any that you previously sent, i.e. via the parent header.
 * Kernel callbacks now uniformly return a value with the shape
   `{ status = "pending" | "done" } | { status = "ready", value = ... }`
+* Type stubs overhauled to be slightly more principled
 
 CLI
 * Added cursed heuristics to determine which error information to use from